+++ src/libtools/edge.cc
@@ -27,8 +27,8 @@ namespace tools {
 
 using namespace std;
 
-Edge::Edge(int ID1, int ID2) {
-  vertices_ = vector<int>{min({ID1, ID2}), max({ID1, ID2})};
+Edge::Edge(long int ID1, long int ID2) {
+  vertices_ = vector<long int>{min({ID1, ID2}), max({ID1, ID2})};
 }
 
 int Edge::getOtherEndPoint(int ver) const {
