+++ src/tests/test_graph_bf_visitor.cc
@@ -45,7 +45,7 @@ BOOST_AUTO_TEST_CASE(basic_test) {
   GraphNode gn1;
   GraphNode gn2;
 
-  unordered_map<int, GraphNode> nodes;
+  unordered_map<long int, GraphNode> nodes;
   nodes[0] = gn1;
   nodes[1] = gn2;
 
@@ -115,7 +115,7 @@ BOOST_AUTO_TEST_CASE(basic_test2) {
   GraphNode gn4;
   GraphNode gn5;
 
-  unordered_map<int, GraphNode> nodes;
+  unordered_map<long int, GraphNode> nodes;
   nodes[0] = gn1;
   nodes[1] = gn2;
   nodes[2] = gn3;
